requirements.txt
# local package
-e .

# external requirements
click~=8.0.1
Sphinx~=4.0.2
coverage~=5.5
awscli~=1.19.96
flake8~=3.9.2
python-dotenv>=0.5.1
pandas~=1.2.4
jmespath~=0.10.0
docutils~=0.15.2
lxml~=4.6.3
pip~=21.0.1
wheel~=0.36.2
brotli~=1.0.9
cryptography~=3.4.7
py~=1.10.0
tornado~=6.1
keyring~=22.3.0
rsa~=4.7.2
pyasn1~=0.4.8
pytz~=2021.1
pyyaml~=5.4.1
babel~=2.9.1
typing_extensions~=3.7.4.3
botocore~=1.20.96
colorama~=0.4.3
mock~=4.0.3
python-dateutil~=2.8.1
s3transfer~=0.4.2
ipython~=7.22.0
setuptools~=57.0.0
pyflakes~=2.3.1
importlib_metadata~=3.10.0
markupsafe~=2.0.1
packaging~=20.9
jinja2~=3.0.1
imagesize~=1.2.0
pillow~=8.2.0
requests~=2.25.1
snowballstemmer~=2.1.0
pytest~=6.2.3
pygments~=2.9.0
certifi~=2020.12.5
chardet~=4.0.0
urllib3~=1.26.5
six~=1.16.0
toml~=0.10.2
idna~=2.10
alabaster~=0.7.12
pyparsing~=2.4.7
cython~=0.29.23
future~=0.18.2
dataset~=1.1.2
appscript~=1.1.2